Metadata refactored into base and derived (via crtp) [skip ci]
