+++ src/include/index/index_defs.h
@@ -110,15 +110,13 @@ using StorageFormat =
      }},
     {"0.3",
      {
+         //--------------------------------------------------------------------------------
+         // Arrays to support ivf_flat indexes
          // From external/data/pytest-170/test_ivf_flat_ingestion_f320/array
          // (generated by test/test_ingestion.py on 2023-12-01)
          // Arrays comprising the index (note that temp_data has these as well)
-         {"centroids_array_name", "partition_centroids"},
-         {"index_array_name", "partition_indexes"},
-         {"ids_array_name", "shuffled_vector_ids"},
-         {"parts_array_name", "shuffled_vectors"},
 
-         // temp_data is a group, not an array
+         // @note: temp_data is a group, not an array
          {"partial_write_array_name", "temp_data"},
 
          // In storage_formats.py but not part of group?
@@ -126,8 +124,24 @@ using StorageFormat =
          {"external_ids_array_name", "external_ids"},
          {"updates_array_name", "updates"},
 
+         //--------------------------------------------------------------------------------
+         // Support for vamana indexes
+         // The feature vectors used to build the index
+         // @todo Is this the same as input_vectors?
+         {"feature_vector_array_name", "feature_vectors"},
+
+         // A very small array used to store the medoid
+         {"medoid_array_name", "medoid"},
+
+         // arrays to support the graph -- stored essentially as CSR
+         // @todo Reuse the ivf_index names for ids and indexes?
+         {"adjacency_scores_array_name", "adjacency_scores"},
+         {"adjacency_ids_array_name", "adjacency_ids"},
+         {"adjacency_index_array_name", "adjacency_index"},
+
+         //--------------------------------------------------------------------------------
          // Not arrays
-         // @todo Where are these stored?
+         // @todo Should these be stored?
          {"default_attr_filters", "zstd"},
          {"support_timetravel", "true"},
      }},
@@ -152,7 +166,7 @@ using StorageFormat =
              "base_sizes",            // (json) list
              "dataset_type",          // "vector_search"
              "dtype",                 // "float32", etc (Python dtype names)
-             "index_type",            // "FLAT", "IVF_FLAT"
+             "index_type",            // "FLAT", "IVF_FLAT", "VAMANA"
              "ingestion_timestamps",  // (json) list
              "partition_history",     // (json) list
              "storage_version",       // "0.3"
@@ -173,10 +187,17 @@ using StorageFormat =
              // probably keep with a timestamp if kept?
 
              // For execution of kmeans -- could be useful for debugging and
-             // reproducibility -- need to store as json list
+             // reproducibility -- need to store as json list (because?)
              "max_iter",        // uint64_t -> json list
              "tol",             // float -> json list
              "reassign_ratio",  // float -> json list
+
+             // Similarly for execution of vamana, we have these parameters
+             "L_build",       // uint64_t -> json list
+             "R_max_degree",  // uint64_t -> json list
+             "B_backtrack",   // uint64_t -> json list
+             "alpha_min",     // float -> json list
+             "alpha_max",     // float -> json list
 
              // For execution of ingestion and queries
              // @todo gather from context
