+++ src/include/test/unit_ivf_flat_group.cc
@@ -33,7 +33,6 @@
 #include <tiledb/tiledb>
 #include <tiledb/group_experimental.h>
 
-#include <filesystem>
 #include <string>
 
 #include "index/ivf_flat_group.h"
@@ -100,26 +99,26 @@ TEST_CASE("ivf_flat_group: constructor", "[ivf_flat_group]") {
   std::reference_wrapper<const dummy_index> bar = foo;
   auto m = bar.get().dimension();
 
-  auto x = ivf_flat_index_group(ctx, group_uri, foo);
-  auto y = ivf_flat_index_group(ctx, group_uri, dummy_index{});
+  auto x = ivf_flat_index_group(dummy_index{}, ctx, group_uri);
+  auto y = ivf_flat_index_group(dummy_index{}, ctx, group_uri);
 }
 
 TEST_CASE("ivf_flat_group: default constructor", "[ivf_flat_group]") {
   tiledb::Context ctx;
-  auto x = ivf_flat_index_group(ctx, group_uri, dummy_index{});
+  auto x = ivf_flat_index_group(dummy_index{}, ctx, group_uri);
   x.dump("Default constructor");
 }
 
 TEST_CASE("ivf_flat_group: read constructor", "[ivf_flat_group]") {
   tiledb::Context ctx;
-  auto x = ivf_flat_index_group(ctx, group_uri, dummy_index{}, TILEDB_READ);
+  auto x = ivf_flat_index_group(dummy_index{}, ctx, group_uri, TILEDB_READ);
   x.dump("Read constructor");
 }
 
 TEST_CASE("ivf_flat_group: read constructor with version", "[ivf_flat_group]") {
   tiledb::Context ctx;
-  auto x = ivf_flat_index_group(
-      ctx, group_uri, dummy_index{}, TILEDB_READ, 0, "0.3");
+  auto x =
+      ivf_flat_index_group(dummy_index{}, ctx, group_uri, TILEDB_READ, 0, "0.3");
   x.dump("Read constructor with version");
 }
 
@@ -129,7 +128,7 @@ TEST_CASE(
   tiledb::Context ctx;
 
   CHECK_THROWS_WITH(
-      ivf_flat_index_group(ctx, "I dont exist", dummy_index{}),
+      ivf_flat_index_group(dummy_index{}, ctx, "I dont exist"),
       "Group uri I dont exist does not exist.");
 }
 
@@ -143,7 +142,7 @@ TEST_CASE("ivf_flat_group: write constructor - create", "[ivf_flat_group]") {
   }
 
   ivf_flat_index_group x =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
   x.dump("Write constructor - create");
 }
 
@@ -158,11 +157,11 @@ TEST_CASE(
   }
 
   ivf_flat_index_group x =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
   x.dump("Write constructor - create before open");
 
   ivf_flat_index_group y =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
   x.dump("Write constructor - open");
 }
 
@@ -177,11 +176,11 @@ TEST_CASE(
   }
 
   ivf_flat_index_group x =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
   x.dump("Write constructor - create before open");
 
   ivf_flat_index_group y =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
   x.dump("Write constructor - open for read");
 }
 
@@ -197,15 +196,15 @@ TEST_CASE(
   }
 
   ivf_flat_index_group x =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
   x.dump("Write constructor - create before open");
 
   ivf_flat_index_group y =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
   x.dump("Write constructor - open for write");
 
   ivf_flat_index_group z =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
   x.dump("Write constructor - open for read");
 }
 
@@ -229,28 +228,28 @@ TEST_CASE(
   size_t offset = 0;
 
   ivf_flat_index_group x =
-      ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
 
   SECTION("Just set") {
     SECTION("After create") {
     }
 
     SECTION("After create and read") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
     }
 
     SECTION("After create and write") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
     }
 
     SECTION("After create and write and read") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
     }
 
     SECTION("After create and read and write") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
     }
 
     x.set_ingestion_timestamp(expected_ingestion);
@@ -265,21 +264,21 @@ TEST_CASE(
     }
 
     SECTION("After create and read") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
     }
 
     SECTION("After create and write") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
     }
 
     SECTION("After create and write and read") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
     }
 
     SECTION("After create and read and write") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
     }
 
     x.append_ingestion_timestamp(expected_ingestion);
@@ -294,21 +293,21 @@ TEST_CASE(
     }
 
     SECTION("After create and read") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
     }
 
     SECTION("After create and write") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
     }
 
     SECTION("After create and write and read") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
     }
 
     SECTION("After create and read and write") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
     }
 
     x.set_ingestion_timestamp(expected_ingestion);
@@ -337,21 +336,21 @@ TEST_CASE(
     }
 
     SECTION("After create and read") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
     }
 
     SECTION("After create and write") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
     }
 
     SECTION("After create and write and read") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
     }
 
     SECTION("After create and read and write") {
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_READ);
-      x = ivf_flat_index_group(ctx, tmp_uri, dummy_index{}, TILEDB_WRITE);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_READ);
+      x = ivf_flat_index_group(dummy_index{}, ctx, tmp_uri, TILEDB_WRITE);
     }
 
     x.set_ingestion_timestamp(expected_ingestion);
