+++ src/src/vamana/vamana_index.cc
@@ -39,10 +39,10 @@
 
 #include <cmath>
 #include "detail/graph/nn-graph.h"
-#include "detail/graph/vamana.h"
 #include "detail/linalg/matrix.h"
 #include "detail/linalg/tdb_matrix.h"
 #include "detail/linalg/vector.h"
+#include "index/vamana_index.h"
 
 bool verbose = false;
 bool debug = false;
